refactor: add customHook for handeling of the Players data in Positions
