+++ frontend/src/components/Game/Positions.jsx
@@ -1,12 +1,10 @@
 import {Flex} from '@chakra-ui/react';
 import PropTypes from 'prop-types';
-import {useSelector} from 'react-redux';
 import PlayerIcons from './PlayerIcons';
+import {usePlayers} from '../../customHooks/usePlayers';
 
 export const Positions = ({relativePositionToTable}) => {
-	const game = useSelector((state) => state.game);
-	const players = getPlayers(game.players);
-	const currentPlayerId = game.currentPlayer;
+	const {players, currentPlayerId} = usePlayers();
 
 	if (relativePositionToTable === 0 || relativePositionToTable === 2) {
 		return (
@@ -39,20 +37,6 @@ export const Positions = ({relativePositionToTable}) => {
 
 Positions.propTypes = {
 	relativePositionToTable: PropTypes.number,
-	players: PropTypes.arrayOf(
-		PropTypes.shape({
-			id: PropTypes.number,
-			name: PropTypes.string,
-			position: PropTypes.number,
-			is_alive: PropTypes.bool,
-		}),
-	),
-	currentPlayerId: PropTypes.number,
 };
-
-function getPlayers(players) {
-	const sortedPlayers = [...players]?.sort((a, b) => a.position - b.position);
-	return sortedPlayers;
-}
 
 export default Positions;
