+++ frontend/src/customHooks/usePlayers.js
@@ -1,0 +1,13 @@
+import {useSelector} from 'react-redux/es/hooks/useSelector';
+
+export const usePlayers = () => {
+	const game = useSelector((state) => state.game);
+
+	const getPlayers = (players) => {
+		const sortedPlayers = [...players]?.sort((a, b) => a.position - b.position);
+		return sortedPlayers;
+	};
+	const players = getPlayers(game.players);
+	const currentPlayerId = game.currentPlayer;
+	return {players, currentPlayerId};
+};
